Correct CLZODecompressor::m_HeaderFlags's type to unsigned int, which is *not* necessarily the same as size_t.
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@33718 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ include/util/compress/lzo.hpp
@@ -590,8 +590,8 @@ private:
 
     // Parameters read from header (used for compression).
     // See fStreamFormat flag description.
-    size_t  m_HeaderLen;       ///< Length of the header.
-    size_t  m_HeaderFlags;     ///< Flags used for compression.
+    size_t       m_HeaderLen;   ///< Length of the header.
+    unsigned int m_HeaderFlags; ///< Flags used for compression.
 };
